factor out common code for the nestedrelationmutactions into a base class
